refactor: move jsonSchemaToTypeSpecIR function
